[mod/#5] 구글로그인 서버 통신 함수, 모듈 context 변경
