+++ app/src/main/java/org/sopt/and/di/DataSourceModule.kt
@@ -8,8 +8,6 @@ import org.sopt.and.data.remote.local.LocalHomeImageDataSource
 import org.sopt.and.data.remote.local.LocalHomeImageDataSourceImpl
 import org.sopt.and.data.remote.source.auth.AuthDataSource
 import org.sopt.and.data.remote.source.auth.AuthDataSourceImpl
-import org.sopt.and.data.remote.source.google.GoogleSignInDataSource
-import org.sopt.and.data.remote.source.google.GoogleSignInDataSourceImpl
 import org.sopt.and.data.remote.source.user.UserDataSource
 import org.sopt.and.data.remote.source.user.UserDataSourceImpl
 import javax.inject.Singleton
@@ -30,12 +28,5 @@ abstract class DataSourceModule {
     @Binds
     @Singleton
     abstract fun bindLocalHomeImageDataSource(localHomeDataSourceImpl: LocalHomeImageDataSourceImpl): LocalHomeImageDataSource
-
-    // GoogleSignInDataSource 추가
-    @Binds
-    @Singleton
-    abstract fun bindGoogleSignInDataSource(
-        googleSignInDataSourceImpl: GoogleSignInDataSourceImpl
-    ): GoogleSignInDataSource
 
 }
