+++ app/src/main/java/org/sopt/and/di/RepositoryModule.kt
@@ -6,11 +6,9 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import org.sopt.and.data.remote.local.LocalHomeImageRepositoryImpl
 import org.sopt.and.data.remote.source.auth.AuthRepositoryImpl
-import org.sopt.and.data.remote.source.google.GoogleSignInRepositoryImpl
 import org.sopt.and.data.remote.source.user.UserRepositoryImpl
 import org.sopt.and.domain.repository.AuthRepository
 import org.sopt.and.domain.repository.UserRepository
-import org.sopt.and.domain.repository.google.GoogleSignInRepository
 import org.sopt.and.domain.repository.local.LocalHomeImageRepository
 import javax.inject.Singleton
 
@@ -34,12 +32,5 @@ abstract class RepositoryModule {
     abstract fun bindsLocalHomeImageRepository(
         localHomeImageRepositoryImpl: LocalHomeImageRepositoryImpl
     ): LocalHomeImageRepository
-
-    // GoogleSignInRepository 추가
-    @Binds
-    @Singleton
-    abstract fun bindsGoogleSignInRepository(
-        googleSignInRepositoryImpl: GoogleSignInRepositoryImpl
-    ): GoogleSignInRepository
 
 }
