+++ app/src/main/java/org/sopt/and/di/UseCaseModule.kt
@@ -6,11 +6,9 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import org.sopt.and.domain.repository.AuthRepository
 import org.sopt.and.domain.repository.UserRepository
-import org.sopt.and.domain.repository.google.GoogleSignInRepository
 import org.sopt.and.domain.repository.local.LocalHomeImageRepository
 import org.sopt.and.domain.usecase.auth.LoginUserUseCase
 import org.sopt.and.domain.usecase.auth.RegisterUserUseCase
-import org.sopt.and.domain.usecase.google.GoogleLoginUseCase
 import org.sopt.and.domain.usecase.local.GetLocalHomeImageUseCase
 import org.sopt.and.domain.usecase.user.GetMyHobbyUseCase
 import javax.inject.Singleton
@@ -37,12 +35,5 @@ object UseCaseModule {
     @Singleton
     fun provideGetLocalHomeImageUseCase(localHomeImageRepository: LocalHomeImageRepository):
             GetLocalHomeImageUseCase = GetLocalHomeImageUseCase(localHomeImageRepository)
-
-    // GoogleLoginUseCase 추가
-    @Provides
-    @Singleton
-    fun provideGoogleLoginUseCase(
-        googleSignInRepository: GoogleSignInRepository
-    ): GoogleLoginUseCase = GoogleLoginUseCase(googleSignInRepository)
 
 }
